+++ src/main/java/com/cafe/service/CategoryService.java
@@ -20,7 +20,7 @@ public interface CategoryService {
 	 ResponseEntity<List<Category>> getAllCategory(String filterValue);
 
 	ResponseEntity<String> updateCategory(Map<String, String> requestMap);
-   public  TransactionDetails createTransaction(Double amount);
+   public  TransactionDetails createTransaction(Integer amount);
